Refactored playback related things
