+++ ultrasonic/src/main/java/org/moire/ultrasonic/activity/SelectAlbumActivity.java
@@ -1024,7 +1024,7 @@ public class SelectAlbumActivity extends SubsonicTabActivity
 
 		for (MusicDirectory.Entry song : selection)
 		{
-			DownloadFile downloadFile = getDownloadService().forSong(song);
+			DownloadFile downloadFile = downloader.getValue().getDownloadFileForSong(song);
 			if (downloadFile.isWorkDone())
 			{
 				deleteEnabled = true;
