+++ ultrasonic/src/main/java/org/moire/ultrasonic/service/BiConsumer.java
@@ -1,0 +1,6 @@
+package org.moire.ultrasonic.service;
+
+public abstract class BiConsumer<T, U>
+{
+    public abstract void accept(T t, U u);
+}
