+++ ultrasonic/src/main/java/org/moire/ultrasonic/view/SongView.java
@@ -39,8 +39,10 @@ import org.moire.ultrasonic.featureflags.FeatureStorage;
 import org.moire.ultrasonic.service.DownloadFile;
 import org.moire.ultrasonic.service.DownloadService;
 import org.moire.ultrasonic.service.DownloadServiceImpl;
+import org.moire.ultrasonic.service.Downloader;
 import org.moire.ultrasonic.service.MusicService;
 import org.moire.ultrasonic.service.MusicServiceFactory;
+import org.moire.ultrasonic.service.Player;
 import org.moire.ultrasonic.util.Util;
 import org.moire.ultrasonic.util.VideoPlayerType;
 
@@ -82,7 +84,8 @@ public class SongView extends UpdateView implements Checkable
 	private boolean maximized = false;
 	private boolean useFiveStarRating;
 
-	private Lazy<DownloadServiceImpl> downloadServiceImpl = inject(DownloadServiceImpl.class);
+	private Lazy<Downloader> downloader = inject(Downloader.class);
+	protected Lazy<Player> player = inject(Player.class);
 
 	public SongView(Context context)
 	{
@@ -169,7 +172,7 @@ public class SongView extends UpdateView implements Checkable
 
 		this.song = song;
 
-		this.downloadFile = downloadServiceImpl.getValue().forSong(song);
+		this.downloadFile = downloader.getValue().getDownloadFileForSong(song);
 
 		StringBuilder artist = new StringBuilder(60);
 
@@ -320,7 +323,7 @@ public class SongView extends UpdateView implements Checkable
 	{
 		updateBackground();
 
-		downloadFile = downloadServiceImpl.getValue().forSong(this.song);
+		downloadFile = downloader.getValue().getDownloadFileForSong(this.song);
 		File partialFile = downloadFile.getPartialFile();
 
 		if (downloadFile.isWorkDone())
@@ -410,7 +413,7 @@ public class SongView extends UpdateView implements Checkable
 		viewHolder.fiveStar4.setImageDrawable(rating > 3 ? starDrawable : starHollowDrawable);
 		viewHolder.fiveStar5.setImageDrawable(rating > 4 ? starDrawable : starHollowDrawable);
 
-		boolean playing = downloadServiceImpl.getValue().getCurrentPlaying() == downloadFile;
+		boolean playing = player.getValue().currentPlaying == downloadFile;
 
 		if (playing)
 		{
